Create GITHUB
As told, Git is a version control system which tracks the changes when working with computer codes while GitHub is 
a Web-based Git version control repository hosting service. It provides all of the distributed version control and 
source code management (SCM) functionalities of Git while topping it with a few of its own features. It is a heaven
 for the developers where they can store their projects and get connected with like-minded people. You can think it 
as “cloud for codes”.

So basically it is a place to store your identical working directories or repositories. It literally is a hub for 
the Git repositories which you can use, simply by creating a free account on the GitHub. These accounts come with an 
abundant storage space where you can store your repositories and build a proper profile which holds a great value.
 By default the repositories are public i.e., everyone can see your codes but you can make it private as well. So 
if you are a good coder you can upload your own applications and programs and get recognition from others.